Use crossterm instead of termion & update libdwarf demo
